requirements.txt
django
psycopg2-binary
djangorestframework
python-dotenv
pillow
Faker
django-filter
djangorestframework-simplejwt
coverage